requirements.txt
requests
cryptography
python-dotenv
urllib3<2.0